+++ src/extension/model/Extension.ts
@@ -2,11 +2,11 @@ import * as vscode from "vscode";
 import { ContainerFactory } from "../../container/build/ContainerFactory";
 import { ContainerInterface } from "../../container/interface/ContainerInterface";
 import { ConstructorType } from "../../container/type/ConstructorType";
-import { FileObserverInterface } from "../../service/filesystem/file/interface/FileObserverInterface";
+import { FilesystemObserverInterface } from "../../service/filesystem/observer/interface/FilesystemObserverInterface";
 import { FileTypeEnum } from "../../utils/enum/FileTypeEnum";
 import { ExtensionInterface } from "../interface/ExtensionInterface";
 import { FileGenerationCommandRegistry } from "../registry/FileGenerationCommandRegistry";
-import { FileObserverRegistry } from "../registry/FileObserverRegistry";
+import { ObserverRegistry } from "../registry/ObserverRegistry";
 import { FileGenerationCommand } from "./../command/FileGenerationCommand";
 
 /**
@@ -24,6 +24,11 @@ export class Extension implements ExtensionInterface {
     private container: ContainerInterface;
 
     /**
+     * Tracks if observers have been registered
+     */
+    private hasObserversRegistered: boolean | undefined;
+
+    /**
      * Creates a new Extension instance with default container
      */
     constructor() {
@@ -38,7 +43,7 @@ export class Extension implements ExtensionInterface {
     public activate(context: vscode.ExtensionContext): this {
         this.initialize(context);
         this.addFileCreationCommands(context);
-        this.addFileOvservers(context);
+        this.addLazyObserverRegistration(context);
         return this;
     }
 
@@ -83,38 +88,40 @@ export class Extension implements ExtensionInterface {
     }
 
     /**
-     * Registers all file observers with VS Code
+     * Registers all observers with VS Code
      * @param context The VS Code extension context
      */
-    private async addFileOvservers(context: vscode.ExtensionContext): Promise<void> {
-        if (!(await this.hasPhpFilesInWorkspace())) {
-            return;
+    private async addObservers(context: vscode.ExtensionContext): Promise<boolean> {
+        if (this.hasObserversRegistered || !(await this.hasPhpFilesInWorkspace())) {
+            return false;
         }
+        this.hasObserversRegistered = true;
 
-        const observerRegisty = FileObserverRegistry;
+        const observerRegisty = ObserverRegistry;
         for (const [observerName, observer] of Object.entries(observerRegisty)) {
-            this.addFileObserver(context, observerName, observer);
+            this.addObserver(context, observerName, observer);
         }
+        return true;
     }
 
     /**
-     * Registers a single file observer with VS Code
+     * Registers a single observer with VS Code
      * @param context The VS Code extension context
      * @param name The name of the observer
      * @param observer The constructor type of the observer to register
      */
-    private addFileObserver(
+    private addObserver(
         context: vscode.ExtensionContext,
         name: string,
-        observer: ConstructorType<FileObserverInterface>
+        observer: ConstructorType<FilesystemObserverInterface>
     ): void {
         const observerInstance = this.container.get(observer);
         if (!observerInstance) {
-            vscode.window.showErrorMessage(`Observer ${name} not found`);
+            console.error(`Observer \`${name}\` not found`);
             return;
         }
         if (typeof observerInstance.watch !== "function") {
-            vscode.window.showErrorMessage(`Observer ${name} does not implement \`watch\``);
+            console.error(`Observer \`${name}\` does not implement \`watch\``);
             return;
         }
         observerInstance.watch(context);
@@ -136,6 +143,29 @@ export class Extension implements ExtensionInterface {
             console.error("Error checking for PHP files:", error);
             return false;
         }
+    }
+
+    /**
+     * Sets up a lazy registration mechanism for file system observers
+     * Observers are only registered if PHP files are present in the workspace
+     * If no PHP files exist initially, a watcher is set up to detect when PHP files are created
+     * @param context The VS Code extension context
+     */
+    private async addLazyObserverRegistration(context: vscode.ExtensionContext): Promise<void> {
+        const observersRegistered = await this.addObservers(context);
+        if (observersRegistered) {
+            return;
+        }
+
+        const phpFileWatcher = vscode.workspace.createFileSystemWatcher("**/*.php", false, false, true);
+        phpFileWatcher.onDidCreate(async (uri) => {
+            const observersRegistered = await this.addObservers(context);
+            if (observersRegistered) {
+                phpFileWatcher.dispose();
+            }
+        });
+
+        context.subscriptions.push(phpFileWatcher);
     }
 
     /**