+++ src/extension/observer/FileRenamedObserver.ts
@@ -1,18 +1,31 @@
-import { FileRenameOperationTypeEnum } from "../../service/filesystem/file/enum/FileRenameOperationTypeEnum";
-import { NamespaceRefactorService } from "../../service/namespace/model/NamespaceRefactorService";
-import { FileRenameObserverAbstract } from "./FileRenameObserverAbstract";
+import * as vscode from "vscode";
+import { FilesystemObserverOperationEnum } from "../../service/filesystem/observer/enum/FilesystemObserverOperationEnum";
+import { FilesystemObserverResourceEnum } from "../../service/filesystem/observer/enum/FilesystemObserverResourceEnum";
+import { FilesystemObserverEvent } from "../../service/filesystem/observer/event/FilesystemObserverEvent";
+import { getUriFileName } from "../../utils/filesystem/getUriFileName";
+import { ObserverAbstract } from "./ObserverAbstract";
 
-export class FileRenamedObserver extends FileRenameObserverAbstract {
-    /**
-     * Constructor for the FileRenamedObserver class.
-     * @param namespaceRefactorService - The NamespaceRefactorService instance to be used for refactoring namespaces.
-     */
-    constructor(protected readonly namespaceRefactorService: NamespaceRefactorService) {
-        super(
-            namespaceRefactorService,
-            FileRenameOperationTypeEnum.Renamed,
-            "refactorNamespacesOnFileRenamed",
-            'Would you like to update the declaration identifer to "{0}" and update its references?'
+export class FileRenamedObserver extends ObserverAbstract {
+    protected getConfigurationOptionName(): string {
+        return "refactorNamespacesOnFileRenamed";
+    }
+
+    protected async getConfirmationMessage(oldUri: vscode.Uri, newUri: vscode.Uri): Promise<string> {
+        const name = getUriFileName(newUri);
+        return vscode.l10n.t(
+            'Would you like to update the declaration identifer to "{0}" and update its references?',
+            name
         );
+    }
+
+    protected async isValidEvent(event: FilesystemObserverEvent): Promise<boolean> {
+        return (
+            event.resource === FilesystemObserverResourceEnum.File &&
+            event.operation === FilesystemObserverOperationEnum.Renamed
+        );
+    }
+
+    protected async onRefactorAccepted(oldUri: vscode.Uri, newUri: vscode.Uri): Promise<void> {
+        await this.namespaceRefactorService.refactorFileAndReferences(oldUri, newUri);
     }
 }