+++ src/service/filesystem/observer/interface/FilesystemObserverInterface.ts
@@ -1,8 +1,8 @@
 import * as vscode from "vscode";
 
-export interface FileObserverInterface {
+export interface FilesystemObserverInterface {
     /**
-     * Starts observing file operations in the workspace.
+     * Starts observing filesystem operations in the workspace.
      * @param context The VS Code extension context used to register disposables.
      */
     watch(context: vscode.ExtensionContext): void;