+++ src/service/namespace/interface/NamespaceValidatorInterface.ts
@@ -1,0 +1,3 @@
+export interface NamespaceValidatorInterface {
+    validate(toValidate: string): Promise<boolean>;
+}